Session_3/requirements.txt
google-generativeai==0.3.2
pycoingecko==3.1.0
python-telegram-bot==20.7
cryptopanic-client==0.1.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
schedule==1.2.1 